+++ src/main/java/com/everis/gameStore/repository/SessionsRepository.java
@@ -1,7 +1,5 @@
 package com.everis.gameStore.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +9,6 @@ import com.everis.gameStore.domain.model.Sessions;
  * The Interface SessionsRepository.
  */
 @Repository
-public interface SessionsRepository extends JpaRepository<Sessions, BigInteger> {
+public interface SessionsRepository extends JpaRepository<Sessions, Long> {
 
 }
