+++ src/main/java/com/everis/gameStore/repository/TokenRepository.java
@@ -1,7 +1,5 @@
 package com.everis.gameStore.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +9,6 @@ import com.everis.gameStore.domain.model.Token;
  * The Interface TokenRepository.
  */
 @Repository
-public interface TokenRepository extends JpaRepository<Token, BigInteger> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
 }
